+++ lib/proxy/buttonProxy.dart
@@ -119,6 +119,418 @@ extension ButtonStyleProxy on NikuButton {
   set mouseCursor(MouseCursor cursor) =>
       mouseCursorState = NikuState.all(cursor);
 
+  void useBg({
+    Color? base,
+    Color? disabled,
+    Color? dragged,
+    Color? error,
+    Color? focused,
+    Color? hovered,
+    Color? pressed,
+    Color? scrolledUnder,
+    Color? selected,
+  }) {
+    _initButton();
+
+    backgroundColorState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useFg({
+    Color? base,
+    Color? disabled,
+    Color? dragged,
+    Color? error,
+    Color? focused,
+    Color? hovered,
+    Color? pressed,
+    Color? scrolledUnder,
+    Color? selected,
+  }) {
+    _initButton();
+    _initText();
+
+    style!.textStyleState!.all!.color = null;
+
+    foregroundColorState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useColor({
+    Color? base,
+    Color? disabled,
+    Color? dragged,
+    Color? error,
+    Color? focused,
+    Color? hovered,
+    Color? pressed,
+    Color? scrolledUnder,
+    Color? selected,
+  }) {
+    _initButton();
+    _initText();
+
+    style!.textStyleState?.all?.color = null;
+
+    foregroundColorState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useOverlayColor({
+    Color? base,
+    Color? disabled,
+    Color? dragged,
+    Color? error,
+    Color? focused,
+    Color? hovered,
+    Color? pressed,
+    Color? scrolledUnder,
+    Color? selected,
+  }) {
+    _initButton();
+
+    overlayColorState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useSplash({
+    Color? base,
+    Color? disabled,
+    Color? dragged,
+    Color? error,
+    Color? focused,
+    Color? hovered,
+    Color? pressed,
+    Color? scrolledUnder,
+    Color? selected,
+  }) {
+    _initButton();
+
+    overlayColorState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useShadowColor({
+    Color? base,
+    Color? disabled,
+    Color? dragged,
+    Color? error,
+    Color? focused,
+    Color? hovered,
+    Color? pressed,
+    Color? scrolledUnder,
+    Color? selected,
+  }) {
+    _initButton();
+
+    shadowColorState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useElevation({
+    double? base,
+    double? disabled,
+    double? dragged,
+    double? error,
+    double? focused,
+    double? hovered,
+    double? pressed,
+    double? scrolledUnder,
+    double? selected,
+  }) {
+    _initButton();
+
+    elevationState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void usePadding({
+    EdgeInsets? base,
+    EdgeInsets? disabled,
+    EdgeInsets? dragged,
+    EdgeInsets? error,
+    EdgeInsets? focused,
+    EdgeInsets? hovered,
+    EdgeInsets? pressed,
+    EdgeInsets? scrolledUnder,
+    EdgeInsets? selected,
+  }) {
+    _initButton();
+
+    paddingState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useP({
+    NikuEdgeInsets? base,
+    NikuEdgeInsets? disabled,
+    NikuEdgeInsets? dragged,
+    NikuEdgeInsets? error,
+    NikuEdgeInsets? focused,
+    NikuEdgeInsets? hovered,
+    NikuEdgeInsets? pressed,
+    NikuEdgeInsets? scrolledUnder,
+    NikuEdgeInsets? selected,
+  }) {
+    _initButton();
+
+    paddingState = NikuState(
+      disabled: disabled?.value,
+      dragged: dragged?.value,
+      error: error?.value,
+      focused: focused?.value,
+      hovered: hovered?.value,
+      pressed: pressed?.value,
+      scrolledUnder: scrolledUnder?.value,
+      selected: selected?.value,
+      base: base?.value,
+    );
+  }
+
+  void useMinimumSize({
+    Size? base,
+    Size? disabled,
+    Size? dragged,
+    Size? error,
+    Size? focused,
+    Size? hovered,
+    Size? pressed,
+    Size? scrolledUnder,
+    Size? selected,
+  }) {
+    _initButton();
+
+    minimumSizeState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useMaximumSize({
+    Size? base,
+    Size? disabled,
+    Size? dragged,
+    Size? error,
+    Size? focused,
+    Size? hovered,
+    Size? pressed,
+    Size? scrolledUnder,
+    Size? selected,
+  }) {
+    _initButton();
+
+    maximumSizeState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useSide({
+    BorderSide? base,
+    BorderSide? disabled,
+    BorderSide? dragged,
+    BorderSide? error,
+    BorderSide? focused,
+    BorderSide? hovered,
+    BorderSide? pressed,
+    BorderSide? scrolledUnder,
+    BorderSide? selected,
+  }) {
+    _initButton();
+
+    sideState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useShape({
+    OutlinedBorder? base,
+    OutlinedBorder? disabled,
+    OutlinedBorder? dragged,
+    OutlinedBorder? error,
+    OutlinedBorder? focused,
+    OutlinedBorder? hovered,
+    OutlinedBorder? pressed,
+    OutlinedBorder? scrolledUnder,
+    OutlinedBorder? selected,
+  }) {
+    _initButton();
+
+    shapeState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
+  void useRounded({
+    double base = 0,
+    double? disabled,
+    double? dragged,
+    double? error,
+    double? focused,
+    double? hovered,
+    double? pressed,
+    double? scrolledUnder,
+    double? selected,
+  }) {
+    _initButton();
+
+    shapeState = NikuState(
+      disabled: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(disabled ?? base),
+      ),
+      dragged: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(dragged ?? base),
+      ),
+      error: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(error ?? base),
+      ),
+      focused: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(focused ?? base),
+      ),
+      hovered: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(hovered ?? base),
+      ),
+      pressed: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(pressed ?? base),
+      ),
+      scrolledUnder: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(scrolledUnder ?? base),
+      ),
+      selected: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(selected ?? base),
+      ),
+      base: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(base),
+      ),
+    );
+  }
+
+  void useCursor({
+    MouseCursor? base,
+    MouseCursor? disabled,
+    MouseCursor? dragged,
+    MouseCursor? error,
+    MouseCursor? focused,
+    MouseCursor? hovered,
+    MouseCursor? pressed,
+    MouseCursor? scrolledUnder,
+    MouseCursor? selected,
+  }) {
+    mouseCursorState = NikuState(
+      disabled: disabled,
+      dragged: dragged,
+      error: error,
+      focused: focused,
+      hovered: hovered,
+      pressed: pressed,
+      scrolledUnder: scrolledUnder,
+      selected: selected,
+      base: base,
+    );
+  }
+
   void useTextStyle(NikuTextStyle Function(NikuTextStyle) v) {
     _initButton();
 
@@ -405,12 +817,6 @@ abstract class ButtonStyleTextStyleProxy {
   set package(String? v) {
     _initText();
     style!.textStyleState?.all?.package = v;
-  }
-
-  // * backgroundColorMacro
-  set bg(Color color) {
-    _initText();
-    style!.textStyleState?.all?.backgroundColor = color;
   }
 
   // * fontStyleMacro